requirements.txt
streamlit
langchain
langchain_community
langchain-ollama
python-dotenv
requests
apideck-unify
pypdf
watchdog
transformers